fix for larger datasets

+++ include/Volume.h
@@ -24,7 +24,7 @@ namespace pbnj {
             void attenuateOpacity(float amount);
             void setColorMap(std::vector<float> &map);
             void setOpacityMap(std::vector<float> &map);
-            std::vector<int> getBounds();
+            std::vector<long unsigned int> getBounds();
             OSPVolume asOSPRayObject();
 
             std::string ID;
